assert_is_batch_mode {assertive.reflection}	R Documentation

How is R running?

Description

Tests to see if R is running in batch mode/interactively.

Usage

assert_is_batch_mode(severity = getOption("assertive.severity", "stop"))

assert_is_interactive(severity = getOption("assertive.severity", "stop"))

assert_is_r_slave(severity = getOption("assertive.severity", "stop"))

assert_is_slave_r(severity = getOption("assertive.severity", "stop"))

is_batch_mode()

is_interactive()

is_r_slave()

is_slave_r()

Arguments

severity
How severe should the consequences of the assertion be? Either "stop", "warning", "message", or "none".

Value

is_batch_mode returns TRUE if R is running in batch mode. is_interactive returns TRUE if R is running interactively.

See Also

EnvVar and interactive.

Examples

is_batch_mode()
is_interactive()
is_r_slave()

[Package assertive.reflection version 0.0-4 Index]