assert_is_architect {assertive.reflection}	R Documentation

Are you running R?

Description

Checks to see what type of R you are running.

Usage

assert_is_architect(severity = getOption("assertive.severity", "stop"))

assert_is_emacs(severity = getOption("assertive.severity", "stop"))

assert_is_revo_r(severity = getOption("assertive.severity", "stop"))

assert_is_rstudio(severity = getOption("assertive.severity", "stop"))

assert_is_rstudio_desktop(severity = getOption("assertive.severity", "stop"))

assert_is_rstudio_server(severity = getOption("assertive.severity", "stop"))

assert_is_visual_studio(severity = getOption("assertive.severity", "stop"))

assert_is_r(severity = getOption("assertive.severity", "stop"))

assert_is_r_alpha(severity = getOption("assertive.severity", "stop"))

assert_is_r_beta(severity = getOption("assertive.severity", "stop"))

assert_is_r_devel(severity = getOption("assertive.severity", "stop"))

assert_is_r_patched(severity = getOption("assertive.severity", "stop"))

assert_is_r_release_candidate(severity = getOption("assertive.severity",
  "stop"))

assert_is_r_release(severity = getOption("assertive.severity", "stop"))

assert_is_r_revised(severity = getOption("assertive.severity", "stop"))

assert_is_r_stable(severity = getOption("assertive.severity", "stop"))

is_architect()

is_emacs()

is_revo_r()

is_rstudio()

is_visual_studio()

is_r()

is_r_alpha()

is_r_beta()

is_r_devel()

is_r_patched()

is_r_release_candidate()

is_r_release()

is_r_revised()

is_r_stable()

Arguments

severity
How severe should the consequences of the assertion be? Either "stop", "warning", "message", or "none".

Value

is_r wraps is.R, providing more information on failure. is_r_stable, is_r_patched, is_r_devel, etc., tell you what type of R build you are running. is_architect, is_rstudio and is_revo_r tell you if you are running Architect/StatET, RStudio, or Revolution Analytics' Revolution R build. is_slave_r tells you if you are running a slave instance of R (e.g. when building a package with devtools or using a cluster). The assert_* functions return nothing but throw an error if the corresponding is_* function returns FALSE.

References

http://www.revolutionanalytics.com/revolution-r-open

See Also

is.R, version, isAvailable.

Examples

# If this is FALSE, you really need to ditch that old copy of S-PLUS
is_r()
assertive.base::dont_stop(assert_is_r())
# Release, patched, devel, etc.
is_r_release()
is_r_patched()
is_r_devel()
is_r_alpha()
is_r_beta()
is_r_release_candidate()
is_r_revised()
switch(
  version$status,
  Patched                        = assert_is_r_patched(),
  "Under development (unstable)" = assert_is_r_devel(),
  alpha                          = assert_is_r_alpha(),
  beta                           = assert_is_r_beta(),
  RC                             = assert_is_r_release_candidate(),
  Revised                        = assert_is_r_revised(),
  assert_is_r_release()
)
# IDE
is_architect()
is_emacs()
is_visual_studio()
is_rstudio()
# Custom R distribution
is_revo_r()

[Package assertive.reflection version 0.0-4 Index]