assert_is_r_current {assertive.reflection}	R Documentation

Is this version of R up to date?

Description

Check if this version of R is as new as the current release version of R.

Usage

assert_is_r_current(severity = getOption("assertive.severity", "stop"))

assert_is_current_r(severity = getOption("assertive.severity", "stop"))

is_r_current(cran = getOption("repos", c(CRAN =
  "http://cran.r-project.org"))["CRAN"])

Arguments

severity
How severe should the consequences of the assertion be? Either "stop", "warning", "message", or "none".

cran
A string giving the URL of the CRAN repository to check.

Value

An object of class R_system_version giving the current release version of R.

Note

Development versions of R can have versions higher than the current release version of R. For convenience, these will return TRUE.

Examples

# This example is marked "don't test" since it requires an 
# internet connection and is potentially long running
is_r_current()

[Package assertive.reflection version 0.0-4 Index]