assert_is_rstudio_current {assertive.reflection}	R Documentation

Is RStudio the current version?

Description

Checks to see if the running version of RStudio is the current version.

Usage

assert_is_rstudio_current(severity = getOption("assertive.severity", "stop"))

is_rstudio_current()

Arguments

severity
How severe should the consequences of the assertion be? Either "stop", "warning", "message", or "none".

Value

is_rstudio_current returns TRUE or FALSE, and assert_is_rstudio_current throws an error in the event of an out of date RStudio. Non-RStudio IDEs throw an error.

References

This function is engineered from the downloadUpdateInfo function from https://github.com/rstudio/rstudio/blob/master/src/cpp/session/modules/SessionUpdates.R where the string for the OS is described in beginUpdateCheck from https://github.com/rstudio/rstudio/blob/master/src/cpp/session/modules/SessionUpdates.cpp

See Also

is_rstudio, is_rstudio_desktop

[Package assertive.reflection version 0.0-4 Index]