sys_get_locale {assertive.reflection}	R Documentation

Get or set the system locale

Description

Wrappers to Sys.getlocale and Sys.setlocale for getting and setting the system locale.

Usage

sys_get_locale(simplify = FALSE, remove_empty_categories = TRUE)

sys_set_locale(..., l = list())

Arguments

simplify
If TRUE, the locale settings are returned as a character vector, otherwise a list.

remove_empty_categories
if TRUE, don't include empty categories.

...
Name-value pairs of locale categories to set.

l
A list, as an alternative method of passing local categories to set.

Value

A named list or vector giving the system locale names. sys_set_locale invisibly returns the locale settings *before* making changes (like setwd and options do).

See Also

Sys.getlocale.

Examples

(current_locale <- sys_get_locale())

# Output simplified to character vector
sys_get_locale(simplify = TRUE)
## Not run: 
# Not run since it (temporarily) affects system settings
english <- if(is_windows()) "English.United_Kingdom" else 
  if(is_mac()) "en_GB" else 
  if(is_linux()) "en_GB.utf8" else
  "en"
sys_set_locale(LC_MONETARY = english)
sys_get_locale()
sys_set_locale(l = current_locale)  #restore everything

## End(Not run)

[Package assertive.reflection version 0.0-4 Index]